move json/cuda code to p16.

+++ p16/zaudiocapturethread.cpp
@@ -180,7 +180,7 @@ void ZAudioCaptureThread::run()
         gGblPara.m_bGblRst2Exit=true;
         return;
     }
-    if(request_periods!=periods)
+    if((int)request_periods!=periods)
     {
         qDebug("<Warning>:Audio CapThread,requested %d periods,but recieved %d.\n", request_periods, periods);
     }
